Named pipe path generated based on Mumble PID (Windows)
This commit appends the Mumble process ID to the default name of the pipe
used to communicate with the overlay clients:

Previous name: "MumbleOverlayPipe"
New naming scheme: "MumbleOverlayPipeXXX" <- where XXX is the PID of the
Mumble instance

In turn an Overlay Helper will only communicate with the one client that
actually launched the helper process and not any process, which may have
opened a pipe with said name beforehand.
Only the Windows client is affected from this change.